requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
numpy==2.2.5
numpy-financial==1.0.0
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
